Implemtation of the DataParser class and the parseStudents method.
Implemented the file data parsing method for handling input and creating the necessary student objects based on the file name inputted in the arguments of the Main class.
What is still needed for implementation is error handling, which will be worked on the later today when I have woken from my lack of sleep.

Signed-off-by: Abdon Morales <abdonmorales@my.utexas.edu>