Made changes to UML design to complete Stage 1
Making major the almost complete UML design file from Gaphor. In
addition, to note, I have not implemented StudentGraph as I'm going to
work on it tomorrow due to other priorities. In addition, I will start
working on the state diagram based on the current implementation of
Stage 1; this was also applied to UML diagram.

Signed-off-by: Abdon Morales <abdonmorales@my.utexas.edu>